e2e-tests/pyproject.toml
[project]
name = "dailylogger-e2e-tests"
version = "0.1.0"
description = "AI-driven E2E tests for DailyLogger Tauri app"
requires-python = ">=3.10"
dependencies = [
    "playwright>=1.40.0",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
]

[dependency-groups]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]